% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/catch_data.R
\name{catch_data}
\alias{catch_data}
\title{Get fisheries catch data}
\usage{
catch_data(dows, trace = TRUE, rich.out = FALSE)
}
\arguments{
\item{dows}{numeric vector of lake DOWs}

\item{trace}{logical if progress is printed in console}

\item{rich.out}{logical if species richness is also returned}
}
\value{
a list or something, I don't know
}
\description{
Get fisheries catch data from DNR lake finder website using a set of DOW identifiers
}
\examples{
dows <- c('27013700', '47004900')

catch_data(dows)
}

